+++ R/internal-tx2gene.R
@@ -5,9 +5,10 @@
 #'
 #' @param projectDir Project directory.
 #' @param genomeBuild Genome build.
+#' @param release Ensembl release version.
 #'
 #' @return [data.frame] with unique rownames.
-.tx2gene <- function(projectDir, genomeBuild = NULL) {
+.tx2gene <- function(projectDir, genomeBuild = NULL, release = "current") {
     filePath <- file.path(projectDir, "tx2gene.csv")
     if (file.exists(filePath)) {
         # bcbio tx2gene
@@ -20,6 +21,6 @@
         if (is.null(genomeBuild)) {
             stop("Genome build required for annotable tx2gene")
         }
-        tx2gene(genomeBuild)
+        tx2gene(genomeBuild, release = release)
     }
 }